Factor out public internal classes of AppEngineCloudType.
AppEngineCloudType is too big, and contains classes that should be
scoped at the package level since they are used by other classes in the package.
